Organize minting module into algebras/interpreters/models packaging.  Combine VrfProofAlgebra and LeaderElectionMintingAlgebra.
