+++ consensus/src/main/scala/co/topl/consensus/interpreters/BlockHeaderValidation.scala
@@ -7,7 +7,7 @@ import co.topl.algebras.{ClockAlgebra, Store, UnsafeResource}
 import co.topl.codecs.bytes.tetra.instances._
 import co.topl.codecs.bytes.typeclasses.implicits._
 import co.topl.consensus.algebras._
-import co.topl.consensus.models.{BlockHeaderValidationFailure, BlockHeaderValidationFailures}
+import co.topl.consensus.models.{BlockHeaderValidationFailure, BlockHeaderValidationFailures, VrfArgument}
 import co.topl.crypto.signing.{Ed25519VRF, KesProduct}
 import co.topl.crypto.hash.Blake2b256
 import co.topl.crypto.signing.Ed25519
@@ -126,7 +126,7 @@ object BlockHeaderValidation {
                     ed25519vrf
                       .verify(
                         header.eligibilityCertificate.vrfSig.bytes.data,
-                        LeaderElectionValidation.VrfArgument(expectedEta, header.slot).signableBytes,
+                        VrfArgument(expectedEta, header.slot).signableBytes,
                         header.eligibilityCertificate.vkVRF.bytes.data
                       )
                       .pure[F]
