+++ consensus/src/main/scala/co/topl/consensus/models/VrfConfig.scala
@@ -1,0 +1,5 @@
+package co.topl.consensus.models
+
+import co.topl.models.utility.Ratio
+
+case class VrfConfig(lddCutoff: Int, precision: Int, baselineDifficulty: Ratio, amplitude: Ratio)
